Fixed an array typing bug in InternalNode.
The problem was that when we called the constructor, it would create an
array of children of type InternalNode<K,V> if it was mid-way in the
tree. The problem was that this would no longer allow us to add
LeafNodes as children. The solution was to use
java.reflect.Class.getSuperclass() to construct an array of type Node.