Fixed a bug where subtrees had larger bounds than allowed.
The problem, in more detail, was that we didn't check to see if
inserting a new node inside of another one (mid-way through the tree)
would make that node's (the middle one) upper bound greater than the key
in the grandparent that leads to it. This resulted in a tree that looked
like:
        1 | 5
    1 | 10    5 | 6

The solution is to check the bounds of your grandparent. If you exceed
them by placing yourself in the current parent, you need to move up (at
least) one level.